Potential fix for code scanning alert no. 3248: Multiplication result converted to larger type

+++ include/stb_image_write.h
@@ -888,8 +888,8 @@ static void *stbiw__sbgrowf(void **arr, int increment, int itemsize) {
   int m = *arr ? 2 * stbiw__sbm(*arr) + increment : increment + 1;
   void *p = STBIW_REALLOC_SIZED(
       *arr ? stbiw__sbraw(*arr) : 0,
-      *arr ? (stbiw__sbm(*arr) * itemsize + sizeof(int) * 2) : 0,
-      itemsize * m + sizeof(int) * 2);
+      *arr ? ((size_t)stbiw__sbm(*arr) * (size_t)itemsize + (size_t)sizeof(int) * 2) : 0,
+      (size_t)itemsize * (size_t)m + (size_t)sizeof(int) * 2);
   STBIW_ASSERT(p);
   if (p) {
     if (!*arr)
